backend/pip-requirements.txt
Django==1.9.5
djangorestframework==3.3.3
-e git+https://github.com/gabalese/mondo-python.git@master#egg=mondo-python-0.0.2
ipython==4.1.1
Werkzeug==0.11.3
django-debug-toolbar==1.4
django-extensions==1.6.1
